- clear amount field - re-organized components
